+++ integration/spark-common/src/main/scala/org/apache/carbondata/streaming/StreamSinkFactory.scala
