+++ streaming/src/main/java/org/apache/carbondata/streaming/CarbonStreamInputFormat.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.carbondata.hadoop.streaming;
+package org.apache.carbondata.streaming;
 
 import java.io.IOException;
