+++ streaming/src/main/java/org/apache/carbondata/streaming/CarbonStreamRecordReader.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.carbondata.hadoop.streaming;
+package org.apache.carbondata.streaming;
 
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -78,6 +78,7 @@ import org.apache.spark.sql.execution.vectorized.ColumnarBatch;
 import org.apache.spark.sql.types.CalendarIntervalType;
 import org.apache.spark.sql.types.Decimal;
 import org.apache.spark.sql.types.DecimalType;
+import org.apache.spark.sql.types.StructField;
 import org.apache.spark.sql.types.StructType;
 import org.apache.spark.unsafe.types.CalendarInterval;
 import org.apache.spark.unsafe.types.UTF8String;
@@ -292,7 +293,8 @@ public class CarbonStreamRecordReader extends RecordReader<Void, Object> {
     cache = cacheProvider.createCache(CacheType.FORWARD_DICTIONARY);
     queryTypes = CarbonStreamInputFormat.getComplexDimensions(carbonTable, storageColumns, cache);
 
-    outputSchema = new StructType(DataTypeUtil.getDataTypeConverter().convertCarbonSchemaToSparkSchema(projection));
+    outputSchema = new StructType((StructField[])
+        DataTypeUtil.getDataTypeConverter().convertCarbonSchemaToSparkSchema(projection));
   }
 
   @Override public boolean nextKeyValue() throws IOException, InterruptedException {
