+++ streaming/src/main/java/org/apache/carbondata/streaming/StreamBlockletReader.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.carbondata.hadoop.streaming;
+package org.apache.carbondata.streaming;
 
 import java.io.EOFException;
 import java.io.IOException;
